refactor: reorgnize structure, add new folder
